# backend/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for TTT Project backend",
        "title": "TTT Project API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/videos": {
            "post": {
                "description": "Create a new video (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a new video",
                "parameters": [
                    {
                        "description": "Video data",
                        "name": "video",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.CreateVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.VideoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/videos/{youtubeId}": {
            "put": {
                "description": "Update an existing video (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update a video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YouTube ID",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Video data",
                        "name": "video",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.UpdateVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.VideoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a video by YouTube ID (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YouTube ID",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/videos/{youtubeId}/tags/{tagId}": {
            "post": {
                "description": "Attach a tag to a video (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Attach a tag to a video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YouTube ID",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Tag ID",
                        "name": "tagId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Detach a tag from a video (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Detach a tag from a video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YouTube ID",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Tag ID",
                        "name": "tagId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/tags/{tagId}/videos": {
            "get": {
                "description": "Get paginated list of videos filtered by tag ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get videos by tag ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tag ID",
                        "name": "tagId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.VideoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/videos": {
            "get": {
                "description": "Get paginated list of videos with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "List all videos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by title",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by tag IDs (comma-separated)",
                        "name": "tag_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.VideoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/videos/{youtubeId}": {
            "get": {
                "description": "Get a single video by its YouTube ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get video by YouTube ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YouTube ID",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.VideoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ttt-project_ttt_v5_backend_internal_domain_dto.CreateVideoRequest": {
            "type": "object",
            "required": [
                "youtube_url"
            ],
            "properties": {
                "youtube_url": {
                    "type": "string"
                }
            }
        },
        "ttt-project_ttt_v5_backend_internal_domain_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "ttt-project_ttt_v5_backend_internal_domain_dto.TagResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "ttt-project_ttt_v5_backend_internal_domain_dto.UpdateVideoRequest": {
            "type": "object",
            "properties": {
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ttt-project_ttt_v5_backend_internal_domain_dto.VideoResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ttt-project_ttt_v5_backend_internal_domain_dto.TagResponse"
                    }
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "upload_date": {
                    "type": "string"
                },
                "youtube_id": {
                    "type": "string"
                }
            }
        }
    }
}